% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SUB_Plots_PatternAndFit.R
\name{addFit}
\alias{addFit}
\title{addFit}
\usage{
addFit(data.df, coeff, col = "red", lwd = 2, lty = 1)
}
\arguments{
\item{data.df}{data frame used for the slope calculation that produced the coefficients (using the Year column here)}

\item{coeff}{are the coefficients (median from posterior or MLE values)}

\item{col}{line color, default is "red"}

\item{lwd}{line width, default is 2}
}
\description{
Function to plot time series pattern with minimal labels
}
