% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_applyAltAlgo.R
\name{applyAltAlgo}
\alias{applyAltAlgo}
\title{applyAltAlgo}
\usage{
applyAltAlgo(
  data.df,
  algorithm = "Minimalist",
  id.col = "CU_ID",
  status.labels = c("Red", "RedAmber", "Amber", "AmberGreen", "Green", "None", "NotRed"),
  status.scores = c(5, 4, 3, 2, 1, NA, 2),
  group.var = NULL,
  tracing = FALSE
)
}
\arguments{
\item{data.df}{data file with metric values, with labels in first column (and optionally an IntStatus column that has results from any integrated status assessments that have been done)}

\item{algorithm}{name of an algorithm (one of "Minimalist",  "FancyPants", "CategoricalRealist","SimplyRed","StateOfTheSalmon1","StateOfTheSalmon2", "StateOfTheSalmon3")}

\item{id.col}{column label for unique case identifier (e.g. "CU_ID")}

\item{status.labels}{vector of labels for status categories. default is c("Red","RedAmber","Amber","AmberGreen","Green", "None","NotRed"),}

\item{status.scores}{score for each of the status labels}

\item{group.var}{column name used for grouping summaries (e.g. by species). default is NULL}
}
\description{
this function archives the alternative algorithm versions used in the Methods paper (https://waves-vagues.dfo-mpo.gc.ca/library-bibliotheque/41207890.pdf). Most users do not need this function!
}
\keyword{trend}
