% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_applyRapidStatusTree.R
\name{applyRapidStatusTree}
\alias{applyRapidStatusTree}
\title{applyRapidStatusTree}
\usage{
applyRapidStatusTree(data.df, id.col = "CU_ID", group.var = NULL)
}
\arguments{
\item{data.df}{data file with metric values, with labels in first column (and optionally an IntStatus column that has results from any integrated status assessments that have been done)}

\item{id.col}{column label for unique case identifier (e.g. "CU_ID")}

\item{group.var}{column name used for grouping summaries (e.g. by species). default is NULL}
}
\description{
this function applies the current version of the rapid status decision tree to a data set of retrospective status metrics calculated with the calcMetrics() function
}
\keyword{trend}
