% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_calcMetrics.R
\name{calcMetrics}
\alias{calcMetrics}
\title{calcMetrics}
\usage{
calcMetrics(
  series.in,
  yrs.in,
  gen.in,
  stk.label = "Stock",
  species.label = "Species",
  series.label = "DataVersion",
  slope.specs = list(num.gen = 3, extra.yrs = 0, filter.sides = 1, slope.smooth = TRUE,
    log.transform = TRUE, out.exp = TRUE, na.rm = FALSE),
  avg.specs = list(avg.type = "geomean", recent.excl = FALSE, lt.smooth = TRUE,
    rel.avg.type = "regular", min.lt.yrs = 20, min.perc.yrs = 20),
  metric.bm = list(RelAbd = c(NA, NA), AbsAbd = c(1000, 10000), LongTrend = c(0.5, 0.75),
    PercChange = c(-25, -15), ProbDeclBelowLBM = c(NA, NA), Percentile = c(0.25, 0.5)),
  retro.start = NULL,
  tracing = TRUE
)
}
\arguments{
\item{series.in}{vector with numeric values}

\item{yrs.in}{vector of years corresponding to the values in series.in
(missing yrs need to be included as NA values in series.in)}

\item{gen.in}{integer value with avg generation time}

\item{stk.label}{label to be used in output file (e.g. "EStu")}

\item{species.label}{label to be used in output file (e.g. "Sk")}

\item{series.label}{label to be used in output file   (e.g. "LogSm_TotSpn")}

\item{slope.specs}{list with arguments for the slope function NAME, see details there}

\item{avg.specs}{list with arguments for the individual metric functions, see details there. (SHOULD CHANGE ARG NAME)}

\item{metric.bm}{list with upper and lower benchmarks for each metric}

\item{retro.start}{user-specified year. if NULL, do retrospective for last 9 yrs of the series}

\item{tracing}{if TRUE, print lots of diagnostic/tracking info to the command lines}
}
\description{
Main function that applies all the individual metric functions and produces a summary output
}
\keyword{benchmarks,}
\keyword{metrics,}
\keyword{slope,}
\keyword{trend}
