% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_calcPercChange.R
\name{calcPercChange}
\alias{calcPercChange}
\title{calcPercChange}
\usage{
calcPercChange(
  X,
  gen.in = 4,
  slope.num.gen = 3,
  extra.yrs = 0,
  genmean.smoothing = TRUE,
  log.transform = TRUE,
  out.exp = TRUE,
  tracing = FALSE
)
}
\arguments{
\item{X}{a data frame of raw values with dimensions Years x Stocks. Row labels are years. NAs values are allowed.}

\item{gen.in}{= average generation time}

\item{slope.num.gen}{number of generations over which to calculate the perc change}

\item{extra.yrs}{to handle COSEWIC "extra year". default is 0}

\item{genmean.smoothing}{if TRUE, apply the smoothSeries() function. default is TRUE}

\item{log.transform}{if TRUE, log-transform the time series before calculating the smoothed series. default is TRUE}

\item{out.exp}{if TRUE, take the exponent of the smoothed series for output. default is TRUE}

\item{tracing}{if TRUE, print various diagnostics to the command line. default is FALSE}
}
\description{
This function applies the calculation of percent change in abundance over time to a data frame (Year x Stock) iteratively for each year
of the time-series using only the data prior to that year.
It calculate declines over a specified number of generation and uses an exponential model to estimate decline rate, as per IUCN guidelines.
For a single vector use  calcPercChangeSimple() to calculate percent declines over the entire time period.
}
\keyword{change,}
\keyword{percent}
\keyword{slope}
