% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_calcPercChangeMCMC.R
\name{calcPercChangeMCMC}
\alias{calcPercChangeMCMC}
\title{calcPercChangeMCMC}
\usage{
calcPercChangeMCMC(
  vec.in,
  method = "jags",
  model.in = NULL,
  perc.change.bm = -25,
  na.skip = FALSE,
  out.type = "short",
  mcmc.plots = FALSE,
  convergence.check = FALSE,
  logged = TRUE
)
}
\arguments{
\item{vec.in}{vector with numeric values}

\item{method}{either "jags" (default), "rstanarm", or "rstan". For properties and discussion of strengths/limitations, refer to the \href{https://github.com/SOLV-Code/MetricsCOSEWIC/wiki/1-Probability-of-Decline:-Estimation-Methods}{MetricsCOSEWIC wiki}.}

\item{model.in}{if NULL, use the built in functions for each method: trend.bugs.1() for jags, ETC}

\item{perc.change.bm}{benchmark for Prob(Decl>BM), default = -25}

\item{na.skip}{if TRUE, skip the calculations when vec.in contains any NA}

\item{out.type}{"short" or "long".
"short" gives summary table of posterior plus "PercChange" and "ProbDecl"
"long" also includes the full mcmc samples and the jags output object}

\item{mcmc.plots}{if true, create the standard series of MCMC diagnostic plots
Note that these are printed to the default device
(i.e. need to external wrap the function call inside a pdf /dev.off call)
To get a plot of the model fit, run this function with out.type = "long",
and then use plot.trend.fit().}

\item{convergence.check}{if TRUE, do an automated convergence check}

\item{logged}{if TRUE, then the input is already log-transformed, and the Perc Change is calculated on the exponentiated fitted values. Default is TRUE}
}
\description{
MCMC version of trend metric calculation
This function calculates the percent change in abundances based on an exponential model of population decline, as per IUCN guidelines
It estimates a distribution of percent declines over the period of the time-series, and provides the probability of declines being greater than a specified threhold
}
\keyword{MCMC,}
\keyword{slope,}
\keyword{trend}
