% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_calcPercChangeSimple.R
\name{calcPercChangeSimple}
\alias{calcPercChangeSimple}
\title{calcPercChangeSimple}
\usage{
calcPercChangeSimple(vec.in)
}
\arguments{
\item{vec.in}{a vector of raw values. NAs are possible.}
}
\description{
This function calculates the percent change on a single vector over the entire time period based on an exponential model of declines
The function calcPercChange() applies the same basic calculation but over shorter time periods and to a data frame of Year x Stock iteratively each year, using only data prior to that year
Note: for now this simply replicates the internal function per.change.mod.fast
}
\examples{
calcPercChangeSimple(vec.in = as.vector(Nile))
}
