% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_calcPercentileSimple.R
\name{calcPercentileSimple}
\alias{calcPercentileSimple}
\title{calcPercentileSimple}
\usage{
calcPercentileSimple(
  vec.in,
  gen.in = 4,
  min.perc.yrs = 20,
  avg.type = "geomean",
  tracing = FALSE,
  recent.excl = FALSE
)
}
\arguments{
\item{vec.in}{a vector of values. NA are possible, but will result in NA trend for any recent time window that includes one or more NA}

\item{gen.in}{use the last gen.in values for the "recent" avg}

\item{min.perc.yrs}{return NA if don't have at least this many years to calculate longterm avg}

\item{avg.type}{"mean","geomean", or "median"}

\item{recent.excl}{if TRUE, then don't use the values from the recent gen as part of the LT avg}
}
\description{
this function calculates the percentile for the last generational average in a single vector.
Function structure and arguments match \code{\link{calcLongTermTrendSimple}}.
NOTE: This function only calculates the percentile value. It is the \code{\link{calcMetrics}} wrapper function that compares
this value to user-specified benchmarks and assigns a status category.
Also note the comments re: percentile benchmarks in the documentation for \code{\link{calcMetrics}}.
}
\examples{
calcPercentileSimple(vec.in = as.vector(Nile) ,
gen.in = 4,min.perc.yrs = 20, avg.type = "geomean",
tracing=FALSE,
recent.excl = FALSE)
}
\keyword{trend}
