% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_comparePercChange.R
\name{comparePercChange}
\alias{comparePercChange}
\title{comparePercChange}
\usage{
comparePercChange(
  du.label,
  du.df,
  yrs.window,
  calc.yr,
  samples.out = TRUE,
  plot.pattern = TRUE,
  plot.posteriors = TRUE,
  plot.boxes = TRUE
)
}
\arguments{
\item{du.label}{short label for the the Designatable Unit (DU)}

\item{du.df}{data frame with the DU time series for : 2 columns, first one is "Year", second the abundance.}

\item{yrs.window}{number of years to use for the percent change calculation (e.g. 3 gen +1)}

\item{calc.yr}{year for which the perc change is calculated}

\item{samples.out}{if TRUE, include the posterior samples in the output}

\item{plot.pattern}{if TRUE, create a plot of the time series with alternative slope estimates}

\item{plot.boxes}{if TRUE, create a plot of the posterior distributions (box plots)}

\item{plot.distributions}{if TRUE, create a plot of the posterior distributions (kernel density plot)}
}
\description{
this function calculates 3 alternative versions of the Perc Change metric: 1 deterministic (MLE) and 2 Bayesian (JAGS via rjags, STAN via RStanArm)
}
