% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_exampleData.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Example data frame for multistock time-series of abundances}
\format{
A dataframe with 3 columns of abundances and 40 rows (years), containing an NA
  * "Stock1" - name of first stock
  * "Stock2" - name of second stock
  * "Stock3" - name of third stock
}
\source{
Generated from running `data-raw/SetUpDataFrame.R`.
}
\usage{
exampleData
}
\description{
Example data frame for multistock time-series of abundances
}
\keyword{datasets}
