% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SUB_Plots_PatternAndFit.R
\name{plotBoxes}
\alias{plotBoxes}
\title{plotBoxes}
\usage{
plotBoxes(
  box.df,
  y.lab = "Label",
  ref.lines = list(BM = -25),
  plot.range = NULL
)
}
\arguments{
\item{box.df}{data frame with 5 rows (values for top whisker, top of box, median, bottom of box, bottom whisker) and 1 column for each distribution. Columns labels become axis labels.}

\item{y.lab}{label for the value axis}

\item{ref.lines}{named list, each element is one value for a vertical ref line}

\item{plot.range}{numeric vector of length 2, specifying the plotting range for the variable.}
}
\description{
Function to plot one or more posterior distributions (boxplots)
}
