% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SUB_Plots_PatternAndFit.R
\name{plotDistribution}
\alias{plotDistribution}
\title{plotDistribution}
\usage{
plotDistribution(x.lab, samples, ref.lines, plot.range = NULL)
}
\arguments{
\item{x.lab}{named list, each element is one set of MCMC sample}

\item{ref.lines}{names list, each element is one value for a vertical ref line}

\item{plot.range}{numeric vector of length 2, specifying the plotting range for the variable.}
}
\description{
Function to plot one or more posterior distributions (density plots)
}
