% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SUB_Plots_PatternAndFit.R
\name{plotPattern}
\alias{plotPattern}
\title{plotPattern}
\usage{
plotPattern(
  yrs,
  vals,
  width = 1,
  color = "grey",
  yrs.axis = FALSE,
  vals.axis = FALSE,
  vals.lim = NULL,
  hgrid = FALSE,
  vgrid = FALSE,
  pch.val = 19,
  pch.bg = NULL
)
}
\arguments{
\item{yrs}{numeric vector of years}

\item{vals}{numeric vector of values}

\item{width}{line width, default is 1}

\item{color}{line color, default is "grey"}

\item{yrs.axis}{if TRUE, display the year axis. default is FALSE}

\item{vals.axis}{if true, display the value axis. default is FALSE}

\item{vals.lim}{vector of length 2 with y-axis limits}

\item{hgrid}{if TRUE, display horizontal grid lines. default is FALSE}

\item{vgrid}{if TRUE, display vertical grid lines. default is FALSE}

\item{pch.val}{numeric value for plotting character. default is 19 (solid pt)}

\item{pch.bg}{text string with color for point fill (if pch allows, e.g. 21)}
}
\description{
Function to plot time series pattern with minimal labels
}
