% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_plotSlopeMetricDetails.R
\name{plotSlopeMetricDetails}
\alias{plotSlopeMetricDetails}
\title{plotSlopeMetricDetails}
\usage{
plotSlopeMetricDetails(
  cu.file,
  metrics.results,
  status.results,
  cu.plot,
  year.plot,
  year.range = c(1960, 2030),
  spn.label = "Wild Spawners",
  cu.label = NULL
)
}
\arguments{
\item{cu.file}{a data frame with CU time series. This is the same data file used by calculateMetricsByCU(). See details there.}

\item{metrics.results}{output from  calcMetrics()}

\item{status.results}{output from generateRapidStatus()}

\item{cu.plot}{CU ID for the conservation unit to be plotted}

\item{year.plot}{year for which to plot the slope diagnostics}

\item{year.range}{time window to be used for x-axis in panel 1, showing the time series.}

\item{spn.label}{label to use for spawner axis in he plots. Default is "Wild Spawners"}

\item{cu.label}{Short label to use for the CU. Default is to use CU_ID provided in argument cu.plot}
}
\description{
this function creates a 3-panel diagnostic plot for the slope metric for 1 specified CU for 1 specified year, using outputs from the mectric calculation and status decision tree functions. The WSP-Rapid-Status-WorkedExamples Repo  includes an illustration: https://github.com/SOLV-Code/WSP-Rapid-Status-WorkedExamples/blob/main/CODE/1_Run_MetricsCalcs%26RapidStatus.R
}
\keyword{details}
\keyword{metric}
\keyword{slope}
