% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_plotStatusDashboards.R
\name{plotStatusDashboards}
\alias{plotStatusDashboards}
\title{plotStatusDashboards}
\usage{
plotStatusDashboards(
  cu.info,
  retro.summary.tbl,
  cu.data,
  out.label = "RapidStatusOut",
  out.filepath = "",
  plot.status.timeline = TRUE
)
}
\arguments{
\item{cu.info}{a data frame with specifications for each CU. For details, see help file for calculateMetricsByCU().}

\item{retro.summary.tbl}{data frame which is the $SummaryTable component of the output from generateRapidStatus()}

\item{cu.data}{a data frame with CU time series. This is the same data frame used for the cu.file argument in the call to calculateMetricsbyCU() and is included as the $Data list element in the output from that function. For details, see that help file.}

\item{out.label}{label to use in the filenames for the output}

\item{out.filepath}{folder for storing the output files}

\item{plot.status.timeline}{Default is TRUE. If FALSE, do not plot the timeline of rapid statuses in the bottom panel}
}
\description{
this function takes outputs from generateRapidStatus() and creates 1 png file with standard diagnostics for each CU.
}
\keyword{dashboard,}
\keyword{plot}
