% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_prepDataForRapidStatus.R
\name{prepDataForRapidStatus}
\alias{prepDataForRapidStatus}
\title{prepDataForRapidStatus}
\usage{
prepDataForRapidStatus(
  cu.info,
  metrics.raw,
  gen.avg.src,
  out.label = "MetricsOut",
  out.filepath = ""
)
}
\arguments{
\item{cu.info}{a data frame with specifications for each CU. For details, see help file for calculateMetricsByCU().}

\item{metrics.raw}{data frame with calculated metrics generated by calculateMetricsByCU().}

\item{gen.avg.src}{data frame with calculated generational averages generated by calculateMetricsByCU().}

\item{out.label}{label to use in the filenames for the output}

\item{out.filepath}{folder for storing the output files}
}
\description{
this reorganizes the output from calculateMetricsByCU() into the format required by generateRapidStatus().
}
\keyword{trend}
