% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_SmoothSeries.R
\name{smoothSeries}
\alias{smoothSeries}
\title{smoothSeries}
\usage{
smoothSeries(
  vec.in,
  gen,
  filter.sides = 2,
  log.transform = TRUE,
  out.exp = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{vec.in}{vector with numeric values}

\item{gen}{generation time (i.e. number of years to average)}

\item{filter.sides}{time window for gen avg (1 = past values only, 2= centered)}

\item{out.exp}{if TRUE, then convert the output back from log space)}

\item{na.rm}{if FALSE, then any gen avg including NA values will be NA}

\item{convergence.check}{if TRUE, do an automated convergence check}
}
\description{
function to calculate a smoothed version of an input vector (calculate the running geometric mean of log_e
transformed data)
}
\keyword{average,}
\keyword{log}
\keyword{smoothing,generational}
\keyword{transform}
