% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Trend_JAGS.R
\name{trend.bugs.1}
\alias{trend.bugs.1}
\title{trend.bugs.1}
\usage{
trend.bugs.1()
}
\description{
Default JAGS model for use in calcPercChangeMCMC(). User-specified variations are also possible.
}
\keyword{BUGS,}
\keyword{JAGS,}
\keyword{MCMC}
\keyword{change,}
\keyword{percent}
