% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_calcLongTermTrend.R
\name{calcLongTermTrend}
\alias{calcLongTermTrend}
\title{calcLongTermTrend}
\usage{
calcLongTermTrend(
  X,
  gen.in = 4,
  recent.num.gen = 1,
  extra.yrs = 0,
  min.lt.yrs = 20,
  avg.type = "geomean",
  tracing = FALSE,
  recent.excl = FALSE
)
}
\arguments{
\item{X}{a data frame with Years x Stocks. Row labels are years, no missing years allowed, NA are possible, but will result in NA Trend for}

\item{extra.yrs}{to handle COSEWIC "extra year"}

\item{avg.type}{"mean","geomean", or "median"}

\item{recent.excl}{if TRUE, then don't use the values from the recent gen as part of the LT avg}
}
\description{
this function applies a basic long-term trend to a data frame of Year x Stock,
with various user options (e.g. log transform, gen avg smoothing, time window to use).
For a single vector, use calcLongTermTrendSimple()
}
\keyword{trend}
