% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_calcLongTermTrendSimple.R
\name{calcLongTermTrendSimple}
\alias{calcLongTermTrendSimple}
\title{calcLongTermTrendSimple}
\usage{
calcLongTermTrendSimple(
  vec.in,
  gen.in = 4,
  min.lt.yrs = 20,
  avg.type = "geomean",
  tracing = FALSE,
  recent.excl = FALSE
)
}
\arguments{
\item{vec.in}{a vector of values. NA are possible, but will result in NA trend for any recent time window that includes one or more NA}

\item{gen.in}{use the last gen.in values for the "recent" avg}

\item{min.lt.yrs}{return NA if don't have at least this many years to calculate longterm avg}

\item{avg.type}{"mean","geomean", or "median"}

\item{recent.excl}{if TRUE, then don't use the values from the recent gen as part of the LT avg}
}
\description{
this function just does a basic long-term trend calculation on a single vector.
calcLongTermTrend() applies the same basic calc retrospectively to a data frame of Year x Stock,
}
\examples{
calcLongTermTrendSimple(vec.in = as.vector(Nile) ,
gen.in = 4,min.lt.yrs = 20, avg.type = "geomean",
tracing=FALSE,
recent.excl = FALSE)
}
\keyword{trend}
