% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_calcPercChange.R
\name{calcPercChange}
\alias{calcPercChange}
\title{calcPercChange}
\usage{
calcPercChange(
  X,
  gen.in = 4,
  slope.num.gen = 3,
  extra.yrs = 0,
  genmean.smoothing = TRUE,
  log.transform = TRUE,
  out.exp = FALSE,
  tracing = FALSE
)
}
\arguments{
\item{X}{a data frame with Years x Stocks. Row labels are years, no missing years allowed, NA are possible, but will result in NA for}

\item{gen.in}{= average generation time}

\item{slope.num.gen}{number of generations over which to calculate the perc change}

\item{extra.yrs}{to handle COSEWIC "extra year". default is 0}

\item{genmean.smoothing}{if TRUE, apply the smoothSeries() function. default is TRUE}

\item{log.transform}{if TRUE, log-transform the time series before calculating the smoothed series. default is TRUE}

\item{out.exp}{if TRUE, take the exponent of the smoothed series before calculating the perc change. default is FALSE}

\item{tracing}{if TRUE, print various diagnostics to the command line. default is FALSE}
}
\description{
this function applies a basic deterministic % change calculation to a data frame of Year x Stock,
with various user options (e.g. log transform, gen avg smoothing, time window to use).
For a single vector, use  calcPercChangeSimple()
}
\keyword{change,}
\keyword{percent}
\keyword{slope}
