% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_calcPercChangeSimple.R
\name{calcPercChangeSimple}
\alias{calcPercChangeSimple}
\title{calcPercChangeSimple}
\usage{
calcPercChangeSimple(vec.in)
}
\arguments{
\item{vec.in}{a vector of values. NA are possible, but will result in NA trend (no infill for now)}
}
\description{
this function just does a basic percent change calculation on a single vector.
calcPercChange() applies the same basic calc retrospectively to a data frame of Year x Stock,
Note: for now this simply replicates the internal function per.change.mod.fast
}
\examples{
calcPercChangeSimple(vec.in = as.vector(Nile))
}
